Add files via upload
Major revision. Now, the algorithm chooses the movement along the edge leading to the vertex with the highest value of the objective function. The previous version of the algorithm chose the movement along the edge, which provides the maximum increment of the objective function per unit path. As a result, the new version of the AlEM algorithm solves Klee-Minty cubes in one iteration.