Add files via upload
A new compilation mode is added: #define PP_GRADIENT. If PP_GRADIENT is defined, then the movement is performed along the edge with the maximum gradient of the objective function. Otherwise, the movement is performed along the edge with the maximum value of the objective function at the end vertex.